+++ src/main/java/it/unibo/game/app/api/BoundingBox.java
@@ -8,6 +8,6 @@ import it.unibo.game.Pair;
 public interface BoundingBox {
     enum Side {UP_DOWN, LEFT_RIGHT};
     enum Corner {LEFT_DOWN , LEFT_UP, RIGHT_DOWN, RIGHT_UP};
-    Map<Corner, Pair<Integer, Integer>> getBox();
+    Map<Corner, Pair<Double,Double>> getBox();
     Optional<Side> collideWith(BoundingBox b);
 }
