+++ src/main/java/it/unibo/game/app/model/ball/BonusBall.java
@@ -12,14 +12,14 @@ public class BonusBall extends GameObjectImpl implements MovingObject{
     private int width;
     private int radius;
 
-    public BonusBall(Pair<Integer,Integer> coord, int h, int w, int r) {
+    public BonusBall(Pair<Double,Double> coord, int h, int w, int r) {
         this.height = h;
         this.width = w;
         this.radius = r;
         this.setPos(centerCalculate(coord));        
     }
-    private Pair<Integer,Integer> centerCalculate(Pair<Integer,Integer> c) {
-        return new Pair<Integer,Integer>(c.getX() + (this.height/2), c.getY() + (this.width/2));
+    private Pair<Double,Double> centerCalculate(Pair<Double,Double> c) {
+        return new Pair<Double,Double>(c.getX() + (this.height/2), c.getY() + (this.width/2));
     }
     public int gerRadius() {
         return this.radius;
