+++ src/main/java/it/unibo/game/app/model/round/RoundEasy.java
@@ -7,10 +7,10 @@ import it.unibo.game.app.model.brick.NormalBrick;
 
 public class RoundEasy extends AbstractRound {
 
-    private int startX;
-    private int startY; 
-    private int endX;
-    private int endY;
+    private Double startX;
+    private Double startY; 
+    private Double endX;
+    private Double endY;
     private int numSurpriseBrick = 0;
 
     public RoundEasy(int jump, int numB, int numS, SizeCalculation size) {
@@ -24,8 +24,8 @@ public class RoundEasy extends AbstractRound {
 
     @Override
     public void setPosBrick() {
-        for(int i = startX; i < endX; i = i + this.getSizeCalc().getBrickDim().getX()) {
-            for(int j = startY; j < endY; j = j + this.getSizeCalc().getBrickDim().getY()) {
+        for(Double i = startX; i < endX; i = i + this.getSizeCalc().getBrickDim().getX()) {
+            for(Double j = startY; j < endY; j = j + this.getSizeCalc().getBrickDim().getY()) {
                 NormalBrick brick = new NormalBrick(BrickType.NORMAL, getSizeCalc().getBrickDim().getY(), getSizeCalc().getBrickDim().getX(),1);
                 brick.setPos(new Pair<>(j,i));
                 super.brick.add(brick);
